requirements.txt
Pillow
matplotlib
seaborn
scipy